add prototype test for prioritizing expected low-runtime threads
This is a prototype of a test. It simulates two processes: one with a CPU hog
when woken, and one with a lightweight thread when woken. Everything gets
pinned to a single core. (Maybe I should unpin the signaler and leave just the
signal-ee threads pinned, that might be an interesting variant.)

This tests that the p90 wake latency of the lightweight process is lower than
that of the hog; that is, the hog is penalized in the priority list. This could
either be based on cumulative execution time, a priority heuristic or anything
else.